requirements.txt
fastapi==0.104.1
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.21
passlib==1.7.4
python-jose==3.5.0
bcrypt==5.0.0
cloudinary==1.44.1
mailjet-rest==1.3.4
python-dotenv==1.0.0
psycopg2-binary==2.9.9